i32

+++ src/system.rs
@@ -457,7 +457,9 @@ pub fn flock(path: &str, options: &FlockOptions) -> Result<i32, String> {
     };
 
     let lck = libc::flock {
-        l_type: lock_type,
+        // l_type is c_short on Linux + macOS; F_RDLCK/F_WRLCK are c_int on
+        // Linux, c_short on macOS. Cast to i16 explicitly for cross-build.
+        l_type: lock_type as i16,
         l_whence: libc::SEEK_SET as i16,
         l_start: 0,
         l_len: 0,
@@ -532,7 +534,7 @@ pub fn flock(path: &str, options: &FlockOptions) -> Result<i32, String> {
 #[cfg(unix)]
 pub fn funlock(fd: i32) -> Result<(), String> {
     let lck = libc::flock {
-        l_type: libc::F_UNLCK,
+        l_type: libc::F_UNLCK as i16,
         l_whence: libc::SEEK_SET as i16,
         l_start: 0,
         l_len: 0,
